% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_fcs.R
\name{convert.navios.FCS}
\alias{convert.navios.FCS}
\title{Convert Navios file to FCS}
\usage{
convert.navios.FCS(files = NULL, raw_dir = ".", pattern = NULL,
  fcs_dir, keep_range = TRUE, verbose = 2)
}
\arguments{
\item{files}{optional character vector with filenames.}

\item{raw_dir}{directory where to look for the files.}

\item{pattern}{this argument is passed on to \code{dir}, see details.}

\item{fcs_dir}{a directory name where FCS files are stored.}

\item{keep_range}{logical, TRUE by default. If TRUE, the ranges of the
resulting FCS are scaled to equal to the linear ranges of the first segment
using the $PxE keywords. If FALSE, no range is scaled.}

\item{verbose}{integer, verbosity, 0 being minimal.}
}
\value{
Nothing, but reports every steps of processing.
}
\description{
A Navios LMD file is FCS file that contains two segments of data. In the
first segment the data are coded in the FCS version 2, data being in log
scale. In the second segment, the data are stored in raw mode (20 bits), ie
at full resolution. Some information (description, scales) are not reported
in the second segment and some conversions are needed. This function extracts
the 2nd segment and export it as FCS 3.
}
\examples{
\dontrun{
# Setup the file locations
raw_dir = "raw_files"
fcs_dir = "fcs_files"
# Then convert
convert.navios.FCS(raw_dir, fcs_dir, verbose = 2)
}

}
